chore(perf): Cleanup dead Web Vitals code
