+++ static/app/views/performance/data.tsx
@@ -14,11 +14,6 @@ import {MutableSearch} from 'sentry/utils/tokenizeSearch';
 import {getCurrentTrendParameter} from 'sentry/views/performance/trends/utils';
 
 import {getCurrentLandingDisplay, LandingDisplayField} from './landing/utils';
-import {
-  getVitalDetailTableMehStatusFunction,
-  getVitalDetailTablePoorStatusFunction,
-  vitalNameFromLocation,
-} from './vitalDetail/utils';
 
 export const DEFAULT_STATS_PERIOD = '7d';
 export const DEFAULT_PROJECT_THRESHOLD_METRIC = 'duration';
@@ -765,50 +760,4 @@ export function generatePerformanceEventView(
     default:
       return eventView;
   }
-}
-
-export function generatePerformanceVitalDetailView(
-  location: Location,
-  organization: Organization
-): EventView {
-  const {query} = location;
-
-  const vitalName = vitalNameFromLocation(location);
-
-  const hasStartAndEnd = query.start && query.end;
-  const savedQuery: NewQuery = {
-    id: undefined,
-    name: t('Vitals Performance Details'),
-    query: 'event.type:transaction',
-    projects: [],
-    fields: [
-      'team_key_transaction',
-      'transaction',
-      'project',
-      'count_unique(user)',
-      'count()',
-      `p50(${vitalName})`,
-      `p75(${vitalName})`,
-      `p95(${vitalName})`,
-      getVitalDetailTablePoorStatusFunction(vitalName),
-      getVitalDetailTableMehStatusFunction(vitalName),
-    ],
-    version: 2,
-    yAxis: [`p75(${vitalName})`],
-  };
-
-  if (!query.statsPeriod && !hasStartAndEnd) {
-    savedQuery.range = getDefaultStatsPeriod(organization);
-  }
-  savedQuery.orderby = decodeScalar(query.sort, '-count');
-
-  const searchQuery = decodeScalar(query.query, '');
-  savedQuery.query = prepareQueryForLandingPage(searchQuery, false);
-
-  const eventView = EventView.fromNewQueryWithLocation(savedQuery, location);
-
-  eventView.additionalConditions.addFilterValues('event.type', ['transaction']);
-  eventView.additionalConditions.addFilterValues('has', [vitalName]);
-
-  return eventView;
 }
