--- static/app/views/performance/transactionSummary/transactionVitals/types.tsx
@@ -1,9 +1,0 @@
-export type Point = {
-  x: number;
-  y: number;
-};
-
-export type Rectangle = {
-  point1: Point;
-  point2: Point;
-};
